refactor(app): relocate 'client' to 'src' directory
Move all files and references from 'client' folder to 'src' to
improve project structure. Update paths across the codebase and
configurations to reflect this change.
