+++ packages/app/package.json
@@ -60,7 +60,7 @@
 		"check": "biome check .",
 		"check:fix": "biome check --apply .",
 		"clean": "rm -rf node_modules build coverage pruned .react-router",
-		"dev": "node --env-file=.development.env --experimental-network-inspection server/server.ts",
+		"dev": "node --env-file=.development.env --experimental-network-inspection ./server/server.ts",
 		"format": "biome check --vcs-enabled=true --vcs-use-ignore-file=true  --formatter-enabled=true --linter-enabled=false .",
 		"format:write": "biome check --vcs-enabled=true --vcs-use-ignore-file=true  --formatter-enabled=true --linter-enabled=false --write .",
 		"lint": "biome ci --vcs-enabled=true --vcs-use-ignore-file=true --vcs-root='../../' --formatter-enabled=false --linter-enabled=true --no-errors-on-unmatched .",
