+++ packages/app/src/entry.server.tsx
@@ -6,7 +6,7 @@ import { isbot } from 'isbot'
 import { renderToPipeableStream } from 'react-dom/server'
 import { type AppLoadContext, type EntryContext, ServerRouter } from 'react-router'
 
-import { getEnv, init } from '#root/client/utils/env.server.ts'
+import { getEnv, init } from '#root/src/utils/env.server.ts'
 
 const ABORT_DELAY = 5_000
