+++ packages/app/src/routes/resume/interests.tsx
@@ -2,7 +2,7 @@ import { T } from '@repo/ui/components/typography/typography.tsx'
 import type * as Model from '@suddenly-giovanni/schema-resume'
 import type { ReactElement } from 'react'
 
-import { generateDjb2Hash } from '#root/client/routes/resume/generate-djb2-hash.ts'
+import { generateDjb2Hash } from '#root/src/routes/resume/generate-djb2-hash.ts'
 
 export function Interests({
 	interests,
