+++ packages/app/src/routes/resume/resume.tsx
@@ -4,12 +4,12 @@ import { Effect } from 'effect'
 import type { ReactElement } from 'react'
 import { Link } from 'react-router'
 
-import { config } from '#root/client/config.ts'
-import { Languages } from '#root/client/routes/resume/languages.tsx'
-import { routesRecord } from '#root/client/routes-record.ts'
-import { loaderFunction } from '#root/client/services/index.ts'
-import { ResumeRepository } from '#root/client/services/resume-repository.ts'
 import hero2800wAssetUrl from '#root/content/assets/hero/giovanni_ravalico-profile_color_e4cily_c_scale,w_2800.webp'
+import { config } from '#root/src/config.ts'
+import { Languages } from '#root/src/routes/resume/languages.tsx'
+import { routesRecord } from '#root/src/routes-record.ts'
+import { loaderFunction } from '#root/src/services/index.ts'
+import { ResumeRepository } from '#root/src/services/resume-repository.ts'
 
 import type { Route } from './+types/resume.ts'
 import { Basics } from './basics.tsx'
