+++ packages/app/src/shell/general-error-boundary.tsx
