+++ packages/app/src/shell/header.tsx
@@ -5,9 +5,9 @@ import { useToggle } from '@repo/ui/hooks/use-toggle.tsx'
 import { clsx } from '@repo/ui/lib/utils.ts'
 import type { ReactElement, SyntheticEvent } from 'react'
 
-import { ThemeSwitch } from '#root/client/routes/resources/theme-switch.tsx'
-import { routesRecord } from '#root/client/routes-record.ts'
 import avatarAssetUrl from '#root/content/assets/giovanni_ravalico-profile_bw.webp'
+import { ThemeSwitch } from '#root/src/routes/resources/theme-switch.tsx'
+import { routesRecord } from '#root/src/routes-record.ts'
 
 import { NavLink } from './nav-link.tsx'
