+++ packages/app/tsconfig.json
@@ -17,6 +17,6 @@
 	"files": [],
 	"references": [
 		{ "path": "./server/tsconfig.node.json" },
-		{ "path": "./client/tsconfig.vite.json" }
+		{ "path": "./src/tsconfig.vite.json"}
 	]
 }
